Re-arranging to avoid circular imports, also including helix inputs
